Update minesweeper.lua
This revision was... a bit unusual. At this point in development I had designed the minesweeper board as a floor, with the intention that players physically drop a flag or bomb on a cell in order to act on that cell. However, I was still trying to figure out how the floor should look, and how I could display numbers on it. One idea I had was to place a grid overtop some dark object... and the darkest object I could think of was a massively upscaled ONI van.