Update minesweeper.lua
By this point in time, I had gotten the ONI van offsets correct and found that it didn't look satisfactory in-game. I abandoned the plan in favor of spawning tons of overlapping grids, with the goal of creating a mostly-opaque blue surface. I intended to draw each cell's number by spawning rotated and downscaled blocks to serve as dots.