Update minesweeper.lua
This is the first revision with user interaction implemented via the Fusion Coil system. This also added the "Action Sack" announcement on game start, and it's when I came up with a concrete plan for showing each cell's number, and for recursively revealing cells when the player selected a safe one. The final recursive-reveal function would use an alternate approach to the optimization described in these to-do notes -- one inspired by a late-night idea I had for handling rooks and bishops in Halo Chess.

The "discoing" bug referenced in the to-do notes is a lighting engine bug which occurs when 640 objects or more are in the player's field of view, even if any objects are obstructed from the player's vision. The lighting engine basically dies, causing all items to randomly flicker and flash like panels on a disco ball.