Current revision of Minesweeper
This is the revision of Minesweeper that was used to create my first video of the gametype.

Here, I implemented the recursive reveal function, switching away from a "reveal" flag to an enum for that purpose. Initially I forgot that recursive revealing should stop at cells with adjacent mines, and should only occur when clicking on cells with no adjacent mines; these were changes I had to go back in and make. In (do_recursive_reveal) you can see the recursive functions, (_process_left) and friends, which I used to perform the Rook-/Bishop-inspired optimization I mentioned in a previous commit.

I implemented the victory condition next. Originally it was hardcoded to award one point for every unmined space that the player revealed; I later added scoring options.

The failure condition was the last thing I implemented. I had thought about it and decided that I wanted some sort of ending cutscene; one idea I had was to spawn a line of explosions stretching from the user-selected landmine to the player. That seemed challenging to implement, so I waited a few minutes and eventually thought of spawning a shower of landmines above the player. It took several tests to get the timing right for the cutscene -- and of course, I had to remember to prevent the victory condition code from running if the full board was revealed as the result of a round failure.