Minesweeper
Attempted to implement FFA behaviors: detecting when the active player quits; forcing other players into Monitors.

The game ending is now its own variable, serving as an enum, instead of being part of a larger flags-mask.

Added script stats, and wrote code so that these should allow players to get an even number of turns on the board.

Removed some debugging code.

The round now ends if the round timer runs out.

I'm still trying to make it so that the ending cutscene doesn't hard-cut to black for the active player (a result of forcibly preventing their spawn and deleting the bipeds that are created as a result of that spawn). No luck so far.