Halo Chess+ plans
There's a game-ruining beeping sound playing during tests. The only sound file in the game that matches it is a sound for the Halo 3 Missile Pod (yes, Reach still has many Halo 3 sound assets), but I can't find anything that references the tag. Wrote down some trial-and-error ideas for getting rid of the noise.

Attempted to fix the problem of the round ending immediately upon checkmate (we want to give the winner time to actually kill the king). Haven't tested that yet.