Halo Chess+ and Minesweeper tweaks
Halo Chess+: Testing indicates that turn order within teams works, at least in a 2v0. However, the UX wasn't ideal: if your turn is up but someone else on your team has control of the board, you get no visual feedback as to who it is. I've modified the widget code to fix this but I haven't tested that modification.

Minesweeper: Starting the game with teams enabled was broken if all players were on Blue Team. I believe I've fixed this but I need to test the new code.