Minesweeper
I remembered that I already had a mechanism for reacting to the end of a turn, and went and put that to use. However, it just... isn't working.

I should probably set the script up to play sounds when reacting to that flag, as a debugging mechanism. That's something I could test for in a single-player team match. I still need a multiplayer test to know that it's properly alternating values, but a single-player match will allow me to know whether and when it's running the code that *should* do that.