Halo Chess+ issue fixes
Added code to delete weapons dropped by slain bipeds (so long as they land on the board).

Fixed a bug in how we handle move selection, which would cause your turn to end instantly if you stood on the same board space for two seconds (even if it wasn't a valid move), or if you stepped off of the board.