Halo Chess+ and Minesweeper quit fixes
The player.killer_type_is opcode is not a reliable way to detect when players quit, so both gametypes now use a trigger to manually detect when (active_player) is gone.

Fixed a bug in Halo Chess+ that allowed you to control both teams' pieces if one team was missing.