WIP: Boundary Break
Added a safety system to handle the case of being unable to supply a newly-swapped Monitor with Active Camo.

The Exuberant Witness effect for Monitors has been somewhat refined; if a Monitor is stationary, it looks as good as is possible, but if the Monitor is player-controlled and moved, things fall apart real quick.